requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
chromadb
renumics-spotlight
numpy
pandas
db-sqlite3